Pressing 'Enter' now sends message
Problem:
You could only send a new message by pressing the send message button.

Solution:
Added an event listener on the new message input box that listens for key press. If the key press is the 'Enter' key the message sends as though the button was pressed.

Notes:
I made sure that pressing enter not on the input box did not send the message.