Fixed bug where we could not start a new conversation from bookpage initially after the developement of the new messaging interface
Moved the automatic clicking of the first person from client.js into its own file, selectTopConversation.js. This was in order for us to import one of two javascript files depending on if we have a new conversation. If this conversation is a new conversation, then we import "newConvoFromBookpage.js" if it is not, then we import selectTopConversation.js. We can see this selection happen in views/partials/javascript.hbs. If the bookOwner is passed in, then we assume the conversation is new. I also added a functionality inside of message_routes.js that checks for that same variable, and retrieves their first name to populate the top field of the new conversation, since we won't have the left person-container to retrieve it from.

	modified:   public/js/client.js
	modified:   public/js/newConvoFromBookPage.js
	new file:   public/js/selectTopConversation.js
	modified:   routes/message_routes.js
	modified:   views/conversation_list.hbs
	modified:   views/partials/javascript.hbs